https://github.com/jackdewinter/pymarkdown/issues/596 - refactored inline_helper and inline_processor, and moved inline modules into new inline package
