+++ publish/coverage.json
@@ -6,8 +6,8 @@
         "totalCovered": 3581
     },
     "lineLevel": {
-        "totalMeasured": 14297,
-        "totalCovered": 14297
+        "totalMeasured": 14451,
+        "totalCovered": 14451
     }
 }
