+++ pymarkdown/inline/inline_response.py
@@ -1,9 +1,10 @@
 """
 Module to hold the response from the inline handle_* functions.
 """
+from __future__ import annotations
 
-from dataclasses import dataclass
-from typing import Any, Optional
+from dataclasses import dataclass, field
+from typing import Optional
 
 
 # pylint: disable=too-many-instance-attributes
@@ -16,7 +17,7 @@ class InlineResponse:
     new_string: Optional[str] = None
     new_string_unresolved: Optional[str] = None
     new_index: Optional[int] = None
-    new_tokens: Any = None
+    new_tokens: list = field(default_factory=list)  # type: ignore
     consume_rest_of_line: bool = False
     original_string: Optional[str] = None
     delta_line_number: int = 0
