+++ test/test_markdown_transform_to_gfm.py
@@ -9,6 +9,10 @@ from pymarkdown.markdown_token import (
     MarkdownTokenClass,
 )
 from pymarkdown.transform_gfm.transform_to_gfm import TransformToGfm
+from pymarkdown.transform_gfm.transform_to_gfm_leaf_token_handlers import (
+    TransformToGfmLeafTokenHandlers,
+)
+from pymarkdown.transform_state import TransformState
 
 
 @pytest.mark.gfm
@@ -60,3 +64,22 @@ def test_gfm_bad_end_token():
 
     # Assert
     assert str(captured_exception) == "Markdown token end type bad not supported."
+
+
+def test_transform_to_gfm_leaf_token_handlers_null():
+    """
+    Test to
+    """
+
+    # Arrange
+    output_html = "<html>"
+
+    # Act
+    transformed_html = TransformToGfmLeafTokenHandlers.null(
+        output_html,
+        MarkdownToken("bob", MarkdownTokenClass.SPECIAL),
+        TransformState([]),
+    )
+
+    # Assert
+    assert transformed_html == output_html
